Cast log_level to u32 to fix Windows MSVC compilation
On Windows/MSVC, bindgen generates ggml_log_level as c_int (i32),
while fit_params() expects u32. The cast is safe since log level
constants are small non-negative values.

+++ src/lib.rs
@@ -840,7 +840,7 @@ fn fit_and_load_model(
             &mut cparams,
             &mut margins,
             fit.n_ctx_min,
-            log_level,
+            log_level as u32,
         )
         .map_err(|e| format!("Parameter fitting failed: {e}"))?;
